+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Modification du vocabulaire concernant les transferts au SAE #1552
 - Possibilité de revenir en arrière sur les flux studio sur les étapes non réalisés #1539
 - Possibilité de retamponner les actes et les annexes (flux studio, actes generique et automatique ) #627
+- SAEConnecteur::sendArchive() devient SAEConnecteur:sendSIP()
 
 ## Suppressions